+++ src/views/IndecisionView.vue
@@ -14,28 +14,7 @@
 <script lang="ts" setup>
 import ChatMessages from '@/components/chat/ChatMessages.vue';
 import MessageBox from '@/components/chat/MessageBox.vue';
-import type { ChatMessage } from '@/interfaces/chat-message.interface';
-import { ref } from 'vue';
+import { useChat } from '@/composables/useChat';
 
-const messages = ref<ChatMessage[]>([
-  {
-    id: new Date().getTime(),
-    message: 'Hola mundo',
-    itsMine: true,
-  },
-  {
-    id: new Date().getTime() + 1,
-    message: 'Hola humano',
-    itsMine: false,
-    image: 'https://yesno.wtf/assets/yes/1-af11222d8d4af90bdab8fc447c8cfebf.gif',
-  },
-]);
-
-const onNewMessage = (text: string) => {
-  messages.value.push({
-    id: new Date().getTime(),
-    message: text,
-    itsMine: true,
-  });
-};
+const { messages, onNewMessage } = useChat();
 </script>
